requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pywebview==4.4.1
pytz==2024.1
reportlab>=4.0.0
tuya-device-sharing-sdk==0.2.9
PyQRCode==1.2.1
cryptography>=46.0.0
opencv-python-headless>=4.10.0,<5
pyinstaller==6.3.0
pywin32==306; platform_system=="Windows"
